New and improved input validation
Init msg now must include the code hash encrypted with the tx

Exec msg now must include the expected contract key encrypted with the tx

Queries unchanged for now

Contract key is now stored in the contract db on init
