+++ x/compute/client/utils/utils.go
@@ -9,7 +9,6 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"github.com/enigmampc/SecretNetwork/x/compute/internal/keeper"
 	"io"
 	"io/ioutil"
 	"log"
@@ -20,8 +19,6 @@ import (
 	cosmwasmTypes "github.com/enigmampc/SecretNetwork/go-cosmwasm/types"
 	regtypes "github.com/enigmampc/SecretNetwork/x/registration"
 	ra "github.com/enigmampc/SecretNetwork/x/registration/remote_attestation"
-
-	"github.com/enigmampc/SecretNetwork/x/compute/internal/types"
 
 	"github.com/enigmampc/cosmos-sdk/client/context"
 	"github.com/miscreant/miscreant.go"
@@ -264,30 +261,6 @@ func (ctx WASMContext) DecryptError(errString string, msgType string, nonce []by
 	}
 
 	return stdErr, nil
-}
-
-func GetContractKey(cliCtx context.CLIContext, contractAddr string) ([]byte, error) {
-	route := fmt.Sprintf("custom/%s/%s/%s", types.QuerierRoute, keeper.QueryContractKey, contractAddr)
-	res, _, err := cliCtx.Query(route)
-	if err != nil {
-		return nil, err
-	}
-
-	return res, nil
-}
-
-func GetCodeHash(cliCtx context.CLIContext, codeID string) ([]byte, error) {
-	route := fmt.Sprintf("custom/%s/%s/%s", types.QuerierRoute, keeper.QueryGetCode, codeID)
-	res, _, err := cliCtx.Query(route)
-	if err != nil {
-		return nil, err
-	}
-
-	var codeResp keeper.GetCodeResponse
-
-	err = json.Unmarshal(res, &codeResp)
-
-	return codeResp.DataHash, nil
 }
 
 type ExecuteMsg struct {
